backend/requirements.txt
backend==0.2.4.1
datasets==2.18.0
Flask==2.2.5
Flask_Cors==4.0.0
langchain==0.3.22
langchain_community==0.3.20
langchain_core==0.3.50
langchain_openai==0.3.12
numpy==2.2.4
pandas==2.2.3
python-dotenv==1.1.0
ragas==0.2.14
